+++ src/states/Initial.h
@@ -35,13 +35,14 @@
 
 namespace lipm_walking
 {
-  /** Initial state.
-   *
-   * Check that contacts match foot positions.
+  /** States of the controller's finite state machine.
    *
    */
   namespace states
   {
+    /** Hold posture and check contacts.
+     *
+     */
     struct Initial : State
     {
       /** Start state.
@@ -75,10 +76,10 @@ namespace lipm_walking
       void hideStartStandingButton();
 
     private:
-      bool postureTaskIsActive_;
-      bool postureTaskWasActive_;
-      bool startStandingButton_;
-      bool startStanding_;
+      bool postureTaskIsActive_; /**< Is the posture task active? */
+      bool postureTaskWasActive_; /**< Was the posture task active at previous run()? */
+      bool startStandingButton_; /**< Is the "Start standing" button displayed in the GUI? */
+      bool startStanding_; /**< Has the user clicked on "Start standing"? */
     };
   }
 }