+++ src/states/Standing.cpp
@@ -26,6 +26,7 @@
  */
 
 #include <lipm_walking/utils/clamp.h>
+#include <lipm_walking/utils/world.h>
 
 #include "Standing.h"
 
@@ -231,8 +232,8 @@ namespace lipm_walking
       return;
     }
     leftFootRatio = clamp(leftFootRatio, 0., 1., "Standing target");
-    sva::PTransformd X_0_mid = sva::interpolate(rightFootContact_.anklePose(), leftFootContact_.anklePose(), leftFootRatio);
-    copTarget_ = X_0_mid.translation();
+    sva::PTransformd X_0_lfr = sva::interpolate(rightFootContact_.anklePose(), leftFootContact_.anklePose(), leftFootRatio);
+    copTarget_ = X_0_lfr.translation();
     leftFootRatio_ = leftFootRatio;
   }